requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
FastAPI==0.115.0
uvicorn==0.30.6
PyMySQL==1.1.1
clear